examples/falcon_sqlalchemy/requirements.txt
SQLAlchemy==1.2.7
graphene==2.1
graphene-sqlalchemy==2.0.0
falcon==1.4.1
gunicorn==19.8.1
graphql-core==2.0
pytest==3.6.0
pytest-runner==4.2
